Move AdsAccountSelectControl
