+++ js/src/components/app-select-control/index.scss
@@ -3,3 +3,7 @@
 		margin-bottom: 0;
 	}
 }
+
+.app-select-control--has-single-value {
+	pointer-events: none;
+}
